+++ tinyexplain/__init__.py
@@ -1,1 +1,4 @@
-from . import task_functions, types, utils
+from . import utils
+from . import types
+from . import explain
+from . import score_functions
