+++ tinyexplain/explain/lime.py
@@ -5,7 +5,7 @@ from sklearn.linear_model import Ridge
 from tinygrad import Tensor
 
 from tinyexplain.types import PostProcessingFunction, ScoreFunction, TinyExplainTask, TinygradModel
-from tinyexplain.utils.logging import Logger
+from tinyexplain.logging import Logger
 
 from .explainer import Explainer
